+++ pkg/providers/property/provider.go
@@ -87,6 +87,7 @@ func Provider() *schema.Provider {
 			"akamai_properties":              dataSourceAkamaiProperties(),
 			"akamai_property_products":       dataSourceAkamaiPropertyProducts(),
 			"akamai_property_hostnames":      dataSourceAkamaiPropertyHostnames(),
+			"akamai_property_include":        dataSourceAkamaiPropertyInclude(),
 			"akamai_properties_search":       dataSourcePropertiesSearch(),
 		},
 		ResourcesMap: map[string]*schema.Resource{