+++ pkg/providers/property/testdata/TestDataPropertyInclude/missing_group_id.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+
+data "akamai_property_include" "include" {
+  contract_id = "ctr_1"
+  include_id  = "inc_1"
+}